+++ middleman/config.py
@@ -101,8 +101,10 @@ class Config(object):
 
         self.command_prefix = self._get_cfg(["command_prefix"], default="!c") + " "
 
+        # Middleman specific config
         self.management_room = self._get_cfg(["middleman", "management_room"], required=True)
         self.management_room_id = self.management_room if self.management_room.startswith("!") else None
+        self.anonymise_senders = self._get_cfg(["middleman", "anonymise_senders"], required=False, default=False)
 
     def _get_cfg(
         self, path: List[str], default: Any = None, required: bool = True,