+++ CHANGELOG.md
@@ -6,3 +6,4 @@ Initial version with the following features
 
 * Messages to bot are relayed to management room
 * Management room users can reply by replying to the messages prefixing with `!reply`
+* Sender messages can be configured as anonymous